:recycle: [Controllers] Move some controllers from v1 to v2
